+++ b4n/ui/widgets/pickers/mod.rs
@@ -1,9 +1,13 @@
+pub use self::base::*;
 pub use self::filter::*;
+pub use self::namespace::*;
 pub use self::pattern::*;
 pub use self::patterns_list::*;
 pub use self::search::*;
 
+mod base;
 mod filter;
+mod namespace;
 mod pattern;
 mod patterns_list;
 mod search;